ICD-SIT 5sec에서 15sec 최대값으로 변경.

+++ src/app_task.h
@@ -78,7 +78,7 @@ private:
 	CHIP_ERROR Init();
 	k_timer mTimer;
 
-	static constexpr uint16_t kTemperatureMeasurementIntervalMs = 60000; /* 60 seconds */
+	static constexpr uint32_t kTemperatureMeasurementIntervalMs = 60000; /* 60 seconds */
 	static constexpr uint16_t kTemperatureMeasurementStep = 100; /* 1 degree Celsius */
 
   static constexpr uint16_t kHumidityMeasurementStep = 100; /* 1 percent */
